move related structs together in controller
